Clicking logo should go to snapcraft.io
It's not ideal (normally you'd expect clicking the logo to take you to
the top page of your current domain) but as there's currently no other
link to https://snapcraft.io, this seems like the quickest fix to
providing one.

QA
--

``` bash
./run
```

Go to <http://127.0.0.1:8004>, browse around, click the logo.